make profile base tests and combine control/data operation commands tests
